Try to load shaders in background/worker threads.
I'm using the FeatureDemo D3D12 with the Bistro interior to test things here, and in my own testing it spends about 75% or more of the startup time just in calls to `Shader::init` (which is where we call `D3DCompile`).

The basic idea of the changes here is to try to load shaders in background threads, but to do so transparently, so that the client of the API need not change their actions. Basically, instead of `Shader::create` immediately calling `Shader::init`, it now returns a `Shader` object right away, and then calls `Shader::init` on another thread. When the user asks for something that requires the result of `init` (currently just the API handle), the implementation will wait for that background work to complete (currently I use `std::future` for this).

Importantly, the implementation of `ProgramVersion` doesn't try to request the API handles, so that we can create a program version from a few shaders and return that further up the call chain, all while the actual shader compile is going on.
In principle, an application should be able to call `Program::getActiveVersion()` to kick off the async compilation, and then only block on completion when they go to actually *use* that program version (e.g., as part of a PSO).

The place where this breaks down right now is that most code actually ends up using a program version almost as soon as it is created, so there isn't much opportunity to buffer up background work.
A nice companion fix would be to have the scene renderer support doing a "pre-pass" over the scene to find the program versions it will use, without actually rendering anything. Such a pass would kick off all the compilation required, while avoiding block on the results.

More notes:

- This change doesn't try to do anything clever with respect to the Slang part of compilation. The actual Slang compilation steps don't take very long as the moment, and they are required to get reflection data (which is used immediately), so trying to parallelize that work would require a lot of Falcor changes.

- One gotcha from the above is that the potential speedup here only applies to D3D12 right now, because for Vulkan we let Slang invoke the downstream compiler for us. The right fix there is probably for Slang to do a similar async approach for its downstream compiler invocations.

- The main reason this change isn't ready for prime-time is that it isn't really clear what we should do with any compilation failures that occur. By pushing the HLSL-to-bytecode compilation into a background thread, we create a situation where we don't know that a compile has failed until late in the process, and the current "abort/retry?" dialog box behavior may be hard to implement in that context (since we may not have the context to meaningfully restart/retry things).

- Another code smell here is that I'm just creating a raw `std::thread` for every shader compile. For now the overhead of this doesn't seem too bad (since the compilation itself is way more expensive), but it would be good to have a real thread pool in place (I'm not convinced the current `Falcor::ThreadPool` does what its name implies...).

Further input is welcome. I'd like to get this into shape to commit soon, since shader compilation is a major factor in startup and iteration times.