Fix Vulkan vertex attribute locations when using cross-compiled shaders
Fixes #23

The Vulkan back end has so far assumed that each vertex attribute "semantic" has a fixed location when used in GLSL coded, e.g.:

```c++
```

The intention is that users can write code like the following in GLSL:

```glsl
layout(location= VERTEX_DIFFUSE_COLOR_LOC)
in vec4 diffuseColor;
```

In those cases, Slang will correctly pass through the declaration (and surface it in reflection data), and the Falcor code will populate a `VertexLayout` with that location, and all is well.

However, when an HLSL or Slang vertex shader is used, the user is instead relying on "semantics":

```hlsl
struct MyVertexInput {
    float3 p : POSITION;
    float4 c : DIFFUSE_COLOR;
};
```

In this case the cross-compiled GLSL will only have two vertex inputs, and they will be auto-assigned indices 0 and 1. Thus any code that constructs a `VertexLayout` expecting color at location 7 will lead to problems.

The fix that this change implements is to query the reflection information during PSO construction. For each vertex attribute in the vertex layout, we check to see if the reflection information contains a parameter with the same semantic. If it does, then we use the location from the reflection information, and ignore the location specified in the `VertexLayout` - this leads to the correct behavior when using HLSL shaders. If there is no matching-semantic input, then we rely on the fixed location already specified - this leads to the correct behavior when using GLSL shaders.