Fix type error in NormalMapFiltering shader code
The code is invoking `+=` on a `float3` and a `float4`. The Microsoft compiler seems to allow this code (maybe it warns and we just don't see it), while Slang diagnoses this as an error.

I'm inclined to say that a user with code like this should fix their code rather than expect Slang to accept this kind of implicit casting behavior.