Added param to capture screen to allow using target rather than backbuffer
Testing calls captureScreen() at the end of onFrameRender.
This actually screenshots the previous frame because the target of this frame hasn't yet been blit'd onto the backbuffer.
This means toggling off text for doesn't really work. The text does toggle, but not for the frame actually being screenshotted.
Having fps breaks testing because it seriously messes up image compare