Improve GraphicsProgram::Desc helpers
The `GraphicsProgram::Desc` type is intended to support a "fluent" style of use:

    auto p = GraphicProgram::create(GraphicsProgram::Desc()
        .sourceFile("foo.hlsl").vertexEntryPoint()
                               .fragmentEntryPoint());

To support this, most of the modifier methods return a `GraphicsProgram::Desc&`.
When somebody added support for `warningsAsErrors()` and `dumpIntermediates()` modifiers, they neglected to follow the precedent, and so these require a more cumbersome format:

    auto d = GraphicsProgram::Desc()
        .sourceFile("foo.hlsl").vertexEntryPoint()
                               .fragmentEntryPoint());
    d.warningsAsErrors(true);
    d.dumpIntermediates(trues);
    auto p = GraphicsProgram::create(d);

This change fixes the declarations to allow the convenient syntax.