Add missing `import`
The `ModelViewer.ps.hlsl` was failing to import one of its dependencies, but because it was being compiled in Slang's "no checking" mode, it got away with this and the result Just Worked.

When compiled by the full Slang compiler (which is now required) we get an "undefined identifier" error for the `VertexOut` type, because it is defined in `DefaultVS.slang` which was not imported.