Add a minimal blurb about Slang to the README
The main point here is to let users know that Slang is in the loop, so that they know who to blame when things go wrong. I thought about adding an explicit comment about what to do if/when compilation seems to get confused, but that would give the impression that we expect things to break on a regular basis.