Initial support for "Shader Components"
At a high level, the ability to use shader components means that a programmer can use (global) generic type parameters in their shader code:

```hlsl
// MyProgram.slang
interface IFrobnicator { ... }
struct SimpleFrobnicator : IFrobnicator { ... };

type_param F : IFrobnicator;
ParameterBlock<F> gFrobnicator;

void main(...) { gFrobnicator.frobincate(...); }
```

Then in the host code, they can load `MyProgram.slang`, create a `ProgramVars` from it, etc., but before they render they need to set the `gFrobnicator` parameter block to a parameter block they created manually from a type (such as `SimpleFrobnicator`). At that point the Slang compiler gets invoked to generate final kernel code specialized to the chosen type, by inferring that the `F` type parameter should be set to `SimpleFrobnicator` based on what was boudn to `gFrobnicator`.

Actually achieving this goal involves several big changes to how Falcor handles shader compilation.

The `ProgramVersion` type no longer represents the final compiled shader kernels. Instead, a `ProgramVersion` just represents the input Slang/HLSL code (from the `Program::Desc`) plus any `#define`s that have been set. This is enough to invoke the Slang compiler with code generation disabled, and get back reflection information, which is what the `preprocessAndCreateProgramVersion` function now does. It is still possible to create a `ProgramVars` from just a `ProgramVersion`.

In order to create actual executable kernels, we must combine a specific `ProgramVersion` (which knows the Slang code to use) with a `ProgramVars` (which knows the types bound for various parameter blocks), to create a new type called `ProgramKernels` which simply encapsulates all the compiled kernels for that program specialized to particular types. In the current common case where there are *no* type parameters, there will only be one `ProgramKernels` for each `ProgramVersion`.

The combination of a program version with program vars is handled in `getCSO`, `getGSO`, etc. These function already took the `ProgramVars` as an argument, so it seems like we were heading in this direction anyway.

In order to support specialization based on the types in parameter blocks, `ParameterBlock` now tracks the type name of its contents, along with a "type ID" that is used when doing specialization (to avoid string-based lookups). The `preprocessAndCreateProgramKernels` function creates kernel code for a particular specialization, and we try to share as much of the Slang logic between this and the `ProgramVersion` generation.

Root signatures probably shouldn't have belonged to `ProgramVars` in the first place, but once we have generic type parameters in the mix it becomes clear that the final root signature can't be fully known until we have the concrete type arguments. For example, the type parameter `F` could be filled in with an empty type (no need to allocate a descriptor set at all for `gFrobnicator`), or a type that recursively contains some `ParameterBlocks` (so that `gFrobnicator` could map to *multiple* descriptor sets).

In the new code, the root signature is part of the `ProgramKernels`. This means that the subroutines for binding data from a `ProgramVars` (which also end up binding root signatures) now take the kernels as an argument, both so that they can extract the root signature when binding it, but also so that they can look up the mapping from their parmaeter blocks to the corresponding root indices.

The way that local/global/all reflectors were stored in a `ProgramVersion` made it hard to get that information out at the right time, so the `ProgramReflectors` have been moved into `ProgramVersion`. The feature was only there for RT programs, but the space was already being consumed for all programs anyway.

Mirroring the split into `Program`, `ProgramVersion`, and `ProgramKernels` in the rasterization case, we now have `RtProgram`, `RtPipelineVersion` and `RtPipelineKernels`. Note that the use of "Pipeline" rather than "Program" in the names is meant to be accurate, since an `RtProgram` is currently actually a *container* for `Program`s.

Just as in the rasterization case, an `RtPipelineVersion` is an `RtProgram` specialized to particular `#define`s, while an `RtPipelineKernels` collects specialized kernels based on the types bound in the various `ProgramVars`. The `RtPipelineKernels` thus fills a similar role to the `ProgramList` in the previous code.

The actual shader table storage has been left in the `RtProgramVars`, although the logic to compute the required record size has migrated to `RtPipelineKernels` because it depends on the actual compiled code.

One somewhat complicated contrivance is that we maintain a "global kernels" object to encapsulate the root signature and root parmameter mapping associated with the global parameters. This is fine by the abstraction (a kernels object can hold zero or more compiled kernels), but still a little weird.

In the context of RT, specialization is only supported for type parameters in the local root signature, but extending support to allow type parameters in the global root signature seems like an important step going forward. There is also no support for having different specializations of the same program manifest in an `RtProgramVars`, but that is similarly an important next step.