Added support for double-quotes in ArgList.
Everything that is inside double-quotes is considered a single arg, even if there is a whitespace inside. Also, If a double-quoted arg is not followed by a space, it is considered part of the the same arg.
For example:
"foo bar" and"now with"space
There are 2 args:
"foo bar"
and"now with"space