GBufferRaster: Properly clear the v-buffer
Invalid items (such as when no geometry was found under a pixel) are
identified by checking whether the first component is `UINT_MAX`.

However since the buffer was cleared to 0, it would always report some
geometry even when there was none, causing passes like the
`MegakernelPathTracer` to hallucinate details when looking in the void.